implemented SpinnerWrapper, moved LeafSpinner into SpinnerWrapper folder
