+++ client/src/components/SpinnerWrapper/SpinnerWrapper.js
@@ -3,8 +3,7 @@ import React from "react";
 
 import { Spinner, spinner_configs } from "src/core/Spinner.js";
 
-// eslint-disable-next-line no-restricted-imports
-import { LeafSpinner } from "../LeafSpinner.js";
+import { LeafSpinner } from "./LeafSpinner.js";
 
 export class SpinnerWrapper extends React.Component {
   render() {
